add filter in all modules

+++ src/components/module/paymentManagement/paymentManagementTable/index.tsx
@@ -163,36 +163,41 @@ const PaymentManagementTable: React.FC<IProps> = ({ searchDebounce, args, setArg
       dataIndex: 'actions',
       key: 'actions',
       className: 'text-center',
-      render: (_, record: IPayment) => (
-        <div className="d-flex flex-row">
-          <Tooltip title="View payment" placement="top" trigger="hover">
-            <Button
-              type="text"
-              size="small"
-              className="cta_btn table_cta_btn"
-              icon={<EyeOutlined />}
-              onClick={() => navigate(`${ROUTES.paymentView}/${record?.id}`)}
-            />
-          </Tooltip>
-          {employeeData?.role === EmployeeRoleName.Admin && (
-            <Tooltip title="Delete payment" placement="top" trigger="hover">
-              <CommonModal
-                title="Delete"
-                content="Are you sure delete this payment?"
-                type="confirm"
-                onConfirm={() => handleDeleteModal(record?.id)}
-              >
-                <Button
-                  type="text"
-                  size="small"
-                  className="cta_btn table_cta_btn"
-                  icon={<DeleteOutlined />}
-                />
-              </CommonModal>
+      render: (_, record: IPayment) => {
+        if (record?.deletedAt) {
+          return '-';
+        }
+        return (
+          <div className="d-flex flex-row">
+            <Tooltip title="View payment" placement="top" trigger="hover">
+              <Button
+                type="text"
+                size="small"
+                className="cta_btn table_cta_btn"
+                icon={<EyeOutlined />}
+                onClick={() => navigate(`${ROUTES.paymentView}/${record?.id}`)}
+              />
             </Tooltip>
-          )}
-        </div>
-      )
+            {employeeData?.role === EmployeeRoleName.Admin && (
+              <Tooltip title="Delete payment" placement="top" trigger="hover">
+                <CommonModal
+                  title="Delete"
+                  content="Are you sure delete this payment?"
+                  type="confirm"
+                  onConfirm={() => handleDeleteModal(record?.id)}
+                >
+                  <Button
+                    type="text"
+                    size="small"
+                    className="cta_btn table_cta_btn"
+                    icon={<DeleteOutlined />}
+                  />
+                </CommonModal>
+              </Tooltip>
+            )}
+          </div>
+        );
+      }
     }
   ];
